+++ tests/playwright/archive/issue-ac-matrices/strk-122-settings-print-button.spec.js
@@ -1,4 +1,4 @@
-import { test, expect } from "./helpers/mocks/extended-test.js";
+import { test, expect } from "../../helpers/mocks/extended-test.js";
 
 const SEED_ITEM = {
   uuid: "strk49-test-item-1",